+++ l10n/da.js
@@ -69,6 +69,7 @@ OC.L10N.register(
     "Panoramas" : "Panoramaer",
     "Reports" : "Raporter",
     "New" : "Ny",
+    "New group" : "Ny gruppe",
     "No panorama yet" : "Intet panorama endnu",
     "No report yet" : "Ingen rapport endnu",
     "Overview" : "Overblik",