+++ l10n/sl.js
@@ -68,6 +68,7 @@ OC.L10N.register(
     "Favorites" : "Priljubljeno",
     "Reports" : "Poročila",
     "New" : "Novo",
+    "New group" : "Nova skupina",
     "No report yet" : "Ni še pripravljenih poročil",
     "Overview" : "Splošni pregled",
     "Remove from favorites" : "Odstrani iz priljubljenih",