+++ l10n/sr.js
@@ -84,6 +84,7 @@ OC.L10N.register(
     "Reports" : "Извештај",
     "Datasets" : "Скупови података",
     "New" : "Нов",
+    "New group" : "Нова група",
     "No panorama yet" : "Још увек нема панораме",
     "No report yet" : "Још увек нема извештаја",
     "Overview" : "Преглед",