+++ l10n/sv.js
@@ -84,6 +84,7 @@ OC.L10N.register(
     "Reports" : "Rapporter",
     "Datasets" : "Datamängder",
     "New" : "Ny",
+    "New group" : "Ny grupp",
     "No panorama yet" : "Ingen panorama än",
     "No report yet" : "Ingen rapport än",
     "Overview" : "Översikt",