+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Threshold lines in charts using Chart.js annotation plugin
 - Show installed version in settings
+- Inline rename for all navigation items
 
 ## 5.7.1 - 2025-07-15
 ### Fixed